+++ tests/Enums/PdfLayoutTest.php
@@ -20,6 +20,9 @@ use PHPUnit\Framework\TestCase;
 
 class PdfLayoutTest extends TestCase
 {
+    /**
+     * @psalm-return \Generator<int, array{0: PdfLayout, 1: PdfVersion}>
+     */
     public static function getVersions(): \Generator
     {
         yield [PdfLayout::DEFAULT, PdfVersion::VERSION_1_3];