+++ tests/PdfDocPropertiesTest.php
@@ -36,6 +36,9 @@ use PHPUnit\Framework\Attributes\DataProvider;
 
 class PdfDocPropertiesTest extends AbstractPdfDocTestCase
 {
+    /**
+     * @psalm-return \Generator<int, array{0: string|int|null, 1: bool}>
+     */
     public static function getIsLinks(): \Generator
     {
         yield [null, false];